tests: Pull buffering tests out of test_lower_exprs
